Added an autoconf macro to check for compiler support for certain options.  Added back warnings and also -fno-enforce-eh-specs.  Fixed all the warnings that slipped into new code.

+++ src/services/http/OW_LocalAuthentication.cpp
@@ -319,7 +319,7 @@ LocalAuthentication::generateNewCookieFile(const String& uid, String& cookieFile
 
 		
 	// Write the servers random number to the temp file
-	if(::write(authfd, cookie.c_str(), cookie.length()) != cookie.length())
+	if(::write(authfd, cookie.c_str(), cookie.length()) != static_cast<ssize_t>(cookie.length()))
 	{
 		int lerrno = errno;
 		::close(authfd);
